pyproject.toml
[tool.poetry]
name = "emsight"
version = "0.1.0"
description = ""
authors = ["wittenator <wittenator@yahoo.de>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.45.0"
sqlalchemy = "^1.3"
bert-serving-client = "^1.10"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
